修正并增加 2，3 维 根据坐标的重心轨迹来求解其坐标集 算法 罗瑶光 20200414

+++ src/org/tinos/deta/demension/Position2D.java
@@ -2,7 +2,7 @@ package org.tinos.deta.demension;
 //Theory: Eclid
 //Application: Yaoguang.luo
 public class Position2D{
-	public Position2D(int x, int y) {
+	public Position2D(double x, double y) {
 		this.setX(x);
 		this.setY(y);
 	}
